item id i32 to i64
Former-commit-id: 875aedca558858991181f0d55d08b4886e31e11b

+++ core/domain/interface/item/sku_service.go
@@ -29,11 +29,11 @@ type (
 	// 商品SKU
 	Sku struct {
 		// 编号
-		ID int32 `db:"id" pk:"yes" auto:"yes"`
+		ID int64 `db:"id" pk:"yes" auto:"yes"`
 		// 产品编号
-		ProductId int32 `db:"product_id"`
+		ProductId int64 `db:"product_id"`
 		// 商品编号
-		ItemId int32 `db:"item_id"`
+		ItemId int64 `db:"item_id"`
 		// 标题
 		Title string `db:"title"`
 		// 图片
