+++ core/domain/order/normal_order.go
@@ -807,7 +807,7 @@ func (o *normalOrderImpl) OnlinePaymentTradeFinish() (err error) {
 }
 
 // 扣减商品库存
-func (o *normalOrderImpl) takeGoodsStock(itemId, skuId int32, quantity int32) error {
+func (o *normalOrderImpl) takeGoodsStock(itemId, skuId int64, quantity int32) error {
 	gds := o.goodsRepo.GetItem(itemId)
 	if gds == nil {
 		return item.ErrNoSuchItem
@@ -1471,7 +1471,7 @@ func (o *subOrderImpl) cancelPaymentOrder() error {
 }
 
 // 退回商品
-func (o *subOrderImpl) Return(snapshotId int32, quantity int32) error {
+func (o *subOrderImpl) Return(snapshotId int64, quantity int32) error {
 	for _, v := range o.Items() {
 		if v.SnapshotId == snapshotId {
 			if v.Quantity-v.ReturnQuantity < quantity {
@@ -1486,7 +1486,7 @@ func (o *subOrderImpl) Return(snapshotId int32, quantity int32) error {
 }
 
 // 撤销退回商品
-func (o *subOrderImpl) RevertReturn(snapshotId int32, quantity int32) error {
+func (o *subOrderImpl) RevertReturn(snapshotId int64, quantity int32) error {
 	for _, v := range o.Items() {
 		if v.SnapshotId == snapshotId {
 			if v.ReturnQuantity < quantity {
